Add puppet community module archive
It turns out the kibana4 module needs an archive module. The nanliu version was always better then the camptocamp version, but it's been moved to upstream puppet community. However that doesn't have a forge release yet (pending https://github.com/puppet-community/puppet-archive/issues/75) so we're adding it here with a specific commit.