Add packages from hiera to here also
The base profile loads packages from hiera with a hiera_hash and then uses create_resources to instantiate them. The problem with that approach is that those packages can not have any dependencies on them. In this commit we add some packahge resources straight into the profile so we can depend on them in other places.